Move GUI step inputs into dedicated module
